Updated badgewidget.php to fix image url
Per Mozilla docs, when the image url is relative, the origin url needs to be prepended to the image url.  If the image url is absolute, ( or starts with http... ) than the image property does contain the entire url to the image. 

An example of a badge with relative url is the Blackboard Exemplary Course badge.  The image does not display in the widget.